+++ packages/primeng/schematics/form/schema.ts
@@ -1,0 +1,5 @@
+export interface FormGeneratorSchema {
+  name: string;
+  fields?: string;
+  path?: string;
+}